dpa_offload: Add DPA IPSec Multiple Instances support
Implemented the code needed to support multiple intances of
DPA IPSec. Basically the IPSec was using only one FMAN and
if the boards had multiple FMANs they could not be used.
Thanks to this patch the IPSec is now able to create multiple
intances using as much of FMAN as available.

Signed-off-by: Andrei Varvara <andrei.varvara@freescale.com>
Change-Id: I98f25f45f1949ce9e922837ae5824e2b186916f2
Reviewed-on: http://git.am.freescale.net:8181/9262
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Marian-Cornel Chereji <marian.chereji@freescale.com>
Reviewed-by: Nicolae-Sebastian Grigore <sebastian.grigore@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>