dpaa_eth: minor improvements for oNIC
1. correct kernel 3.12 checkpatch errors
These error are related to the copyright comments at the beginning of each file
and to the usage of seq_puts for non-formatted strings.

2. create specific frame queue initialization function
Now, the frame queues are initialized with the standard DPAA Ethernet function.
With this patch a specific initialization function that does not have CGR or
configurmation queues is created for oNIC.

3. remove unuseful operations from buffer pool probing function
There is no need for physical address of the buffers since the buffers are in
kernel memory space. Also, the fill of the buffer pools is done when the
interface is raised removing the need for a standard seeding callback.

4. add support for mac address modification
If the user tries to modify the mac address he will receive "operation not
supported". With this patch you will be able to add a different mac address.
This is a "virtual" mac address recognized only in the kernel. oNIC does not
have a mac device underneath.

Change-Id: I8e3d884105f2e8f96667b2f83df16761464be6e7
Signed-off-by: Marian Rotariu <marian.rotariu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/10495
Reviewed-by: Madalin-Cristian Bucur <madalin.bucur@freescale.com>
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>