fman: don't redefine NO_IRQ
This triggered a warning on the v3.12 merge.  Redefining standard
kernel symbols is *wrong*.  Don't do it.

FWIW, NO_IRQ is deprecated and should not be used at all.  Use zero.

I tried removing this redefinition entirely, but that caused a build
failure.  I tried adding an #include of the relevant Linux header from
drivers/net/ethernet/freescale/fman/src/inc/types_linux.h which AFAICT
is the place for Linux dependencies, but it still wasn't getting
included so I gave up and defined it to the same exact thing as Linux.

Note that while this warning did not appear on v3.13-rc3 (presumably
due to headers being included in different order), there are lots of
warnings about "fm_port_suspend missing" both before and after moving
to 3.12.  Please fix these.  Warnings should be treated as errors, not
ignored.

Signed-off-by: Scott Wood <scottwood@freescale.com>
Change-Id: I18e8d71ed8a9494547cb0a2b7dca475deb26c030
Reviewed-on: http://git.am.freescale.net:8181/10770
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Kim Phillips <Kim.Phillips@freescale.com>
Reviewed-by: Madalin-Cristian Bucur <madalin.bucur@freescale.com>