asf_gfar: Patch to fix the crash seen in ifup for bsc9132.
ASF uses two TX queues to avoid contention. BSC9132 even though having
controller which can support two groups was having the DTS file written
for one group. To fix this we allow ASF to override the default tx queue
mapping with the "fsl,tx-bit-map" value given in the DT.

This patch resolves the crash seen in case of bsc9132 board
when the network interface is up and trying to xmit the packet.

Signed-off-by: Sandeep Malik <Sandeep.Malik@freescale.com>
---
CQ:ENGR311623
Change-Id: I9d61b0ac73674774f4b674836478048e9a0c8873
Reviewed-on: http://git.am.freescale.net:8181/12091
Reviewed-by: Claudiu Manoil <claudiu.manoil@freescale.com>
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>