powerpc/e6500: hw tablewalk: fix recursive tlb lock on cpu 0
Commit e2929e1e0ab910995090ce0714beab75618db694 "powerpc/e6500: Make
TLB lock recursive" introduced a bug whereby cpu 0 uses the same value
for "lock held" as is used to indicate that the lock is free.  This
means that cpu 1 can acquire the lock whenever it wants, regardless of
whether cpu 0 has it locked, which in turn means we can get duplicate
TLB entries.

Add one to the CPU value to ensure we do not use zero as a "lock held"
value.

Signed-off-by: Scott Wood <scottwood@freescale.com>
Change-Id: I8c8013adc2e153f19d780a3b202c993054feb47f
Reviewed-on: http://git.am.freescale.net:8181/12823
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Mihai Claudiu Caraman <mihai.caraman@freescale.com>
Reviewed-by: Richard Schmitt <richard.schmitt@freescale.com>