Revert "Patch to modify default ICV lengh of SHA2 to 128 bits."
This reverts commit c9b7e8bcef1fb9476e4198c66bc066d1a22acb24.

Horia is right - this issue must be fixed in userspace to avoid inter-kernel version compatibility (e.g., this now makes the SDK kernel not able to run IPSec SHA2 with any other kernel).  If the tools you are using don't support it, either use a tool that does, or fix the tool itself.

Change-Id: I48db232b1e55bc5bd9957667b88fe27e66114546
Reviewed-on: http://git.am.freescale.net:8181/12825
Reviewed-by: Scott Wood <scottwood@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>
Tested-by: Jose Rivera <German.Rivera@freescale.com>