dpa_offload: Aggregate dynamically created HM operations
The DPA classifier driver allows multiple HM operations to be aggregated
in the same HM node only when importing the HM node resources. When
creating the nodes dynamically, the driver does not search for candidates
for aggregation before creating the HM nodes.

This update improves the dpa_classifier driver to actively check for
HM operation aggregation possibilities to certain neighbor HM nodes even
when using the API in dynamic mode.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: I96bb3eb62be81e3d1c3ac83947f32bee230ace38
Reviewed-on: http://git.am.freescale.net:8181/15220
Reviewed-by: Radu-Andrei Bulie <Radu.Bulie@freescale.com>